ShaderCode + refactoring.

+++ src/base/file.cc
@@ -4,20 +4,20 @@
 
 char* File::ReadWholeFile(const char* file_name,
                           const char* root_path,
-                          unsigned* length,
+                          int* length,
                           bool null_terminate) {
   // Determine how big the file is.
   File file;
   if (file.Open(file_name, root_path)) {
-    unsigned size = file.GetSize();
+    int size = file.GetSize();
     if (size) {
       // Allocate a new buffer and add space for a null terminator.
       char* buffer = new char[size + (null_terminate ? 1 : 0)];
 
       // Read all of it.
-      unsigned bytesRead = file.Read(buffer, size);
+      int bytesRead = file.Read(buffer, size);
       if (!bytesRead) {
-        LOG << "Failed to read a buffer of size: " << (int)size << " from file " << file_name;
+        LOG << "Failed to read a buffer of size: " << size << " from file " << file_name;
         delete[] buffer;
         return NULL;
       }
